Sprint 6

+++ shared/src/persistence/dynamo/cases/saveCase.js
@@ -2,12 +2,48 @@ const {
   createRespondentCaseMapping,
   stripInternalKeys,
   stripWorkItems,
+  createMappingRecord,
 } = require('../../awsDynamoPersistence');
 const { syncWorkItems } = require('../../dynamo/workitems/syncWorkItems');
 const { syncDocuments } = require('../../dynamo/documents/syncDocuments');
 
 const client = require('../../dynamodbClientService');
 
+exports.saveVersionedCase = async ({
+  existingVersion,
+  caseToSave,
+  applicationContext,
+}) => {
+  const TABLE = `efcms-${applicationContext.environment.stage}`;
+
+  // used for associating a case to the latest version
+  const currentVersion = existingVersion;
+  const nextVersionToSave = parseInt(currentVersion || '0') + 1;
+
+  // update the current history
+  await client.put({
+    applicationContext,
+    TableName: TABLE,
+    Item: {
+      pk: caseToSave.caseId,
+      sk: '0',
+      ...caseToSave,
+      currentVersion: `${nextVersionToSave}`,
+    },
+  });
+
+  // add a history entry
+  return await client.put({
+    applicationContext,
+    TableName: TABLE,
+    Item: {
+      pk: caseToSave.caseId,
+      sk: `${nextVersionToSave}`,
+      ...caseToSave,
+      currentVersion: `${nextVersionToSave}`,
+    },
+  });
+};
 /**
  * saveCase
  * @param caseToSave
@@ -21,13 +57,33 @@ exports.saveCase = async ({ caseToSave, applicationContext }) => {
     TableName: TABLE,
     Key: {
       pk: caseToSave.caseId,
-      sk: caseToSave.caseId,
+      sk: '0',
     },
   });
 
-  const currentStatus = currentCaseState.status;
+  if (!currentCaseState) {
+    await createMappingRecord({
+      pkId: caseToSave.userId,
+      skId: caseToSave.caseId,
+      type: 'case',
+      applicationContext,
+    });
 
-  if (!currentCaseState.respondent && caseToSave.respondent) {
+    await createMappingRecord({
+      pkId: caseToSave.docketNumber,
+      skId: caseToSave.caseId,
+      type: 'case',
+      applicationContext,
+    });
+  }
+
+  const currentStatus = (currentCaseState || {}).status;
+
+  if (
+    currentCaseState &&
+    !currentCaseState.respondent &&
+    caseToSave.respondent
+  ) {
     await createRespondentCaseMapping({
       applicationContext,
       caseId: caseToSave.caseId,
@@ -47,16 +103,17 @@ exports.saveCase = async ({ caseToSave, applicationContext }) => {
     currentCaseState,
   });
 
-  // if a stipulated decision was uploaded, create a work item entry
   if (currentStatus !== caseToSave.status) {
-    await client.delete({
-      applicationContext,
-      tableName: TABLE,
-      key: {
-        pk: `${currentStatus}|case-status`,
-        sk: caseToSave.caseId,
-      },
-    });
+    if (currentStatus) {
+      await client.delete({
+        applicationContext,
+        tableName: TABLE,
+        key: {
+          pk: `${currentStatus}|case-status`,
+          sk: caseToSave.caseId,
+        },
+      });
+    }
 
     await client.put({
       applicationContext,
@@ -68,18 +125,17 @@ exports.saveCase = async ({ caseToSave, applicationContext }) => {
     });
   }
 
-  const results = await client.put({
+  const results = await exports.saveVersionedCase({
+    caseToSave,
+    existingVersion: (currentCaseState || {}).currentVersion,
     applicationContext,
-    TableName: TABLE,
-    Item: {
-      pk: caseToSave.caseId,
-      sk: caseToSave.caseId,
-      ...caseToSave,
-    },
   });
 
-  return stripWorkItems(
-    stripInternalKeys(results),
-    applicationContext.isAuthorizedForWorkItems(),
-  );
+  return applicationContext.filterCaseMetadata({
+    cases: stripWorkItems(
+      stripInternalKeys(results),
+      applicationContext.isAuthorizedForWorkItems(),
+    ),
+    applicationContext,
+  });
 };
